+++ src/Components/loginAndSignup/CopyrightComponent.tsx
@@ -1,0 +1,14 @@
+import { Link, Typography } from "@mui/material";
+
+export default function Copyright(props: any) {
+  return (
+    <Typography variant="body2" color="text.secondary" align="center" {...props}>
+      {'Copyright © '}
+      <Link color="inherit" href="https://mui.com/">
+        Oleg Borisov
+      </Link>{' '}
+      {new Date().getFullYear()}
+      {'.'}
+    </Typography>
+  );
+}
